adding util project for shared code
