+++ activiti-core-dependencies/pom.xml
@@ -110,6 +110,11 @@
       </dependency>
       <dependency>
         <groupId>org.activiti</groupId>
+        <artifactId>activiti-spring-conformance-util</artifactId>
+        <version>${activiti.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.activiti</groupId>
         <artifactId>activiti-spring-conformance-set0</artifactId>
         <version>${activiti.version}</version>
       </dependency>
