+++ activiti-spring-conformance/activiti-spring-conformance-set0/pom.xml
@@ -16,6 +16,10 @@
 
   <dependencies>
     <dependency>
+      <groupId>org.activiti</groupId>
+      <artifactId>activiti-spring-conformance-util</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.activiti.api</groupId>
       <artifactId>activiti-api-process-runtime</artifactId>
     </dependency>
