+++ activiti-spring-conformance/activiti-spring-conformance-set1/src/test/java/org/activiti/spring/conformance/set1/ConformanceBasicProcessRuntimeTest.java
@@ -8,7 +8,7 @@ import org.activiti.api.process.runtime.events.listener.ProcessRuntimeEventListe
 import org.activiti.api.runtime.shared.events.VariableEventListener;
 import org.activiti.api.runtime.shared.query.Page;
 import org.activiti.api.runtime.shared.query.Pageable;
-import org.activiti.spring.conformance.set1.security.util.SecurityUtil;
+import org.activiti.spring.conformance.util.security.SecurityUtil;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
