+++ activiti-spring-conformance/activiti-spring-conformance-set1/src/test/java/org/activiti/spring/conformance/set1/ConformanceServiceTaskModifyVariableTest.java
@@ -10,7 +10,7 @@ import org.activiti.api.process.model.events.BPMNSequenceFlowTakenEvent;
 import org.activiti.api.process.model.events.ProcessRuntimeEvent;
 import org.activiti.api.process.runtime.ProcessRuntime;
 import org.activiti.api.runtime.shared.NotFoundException;
-import org.activiti.spring.conformance.set1.security.util.SecurityUtil;
+import org.activiti.spring.conformance.util.security.SecurityUtil;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -18,7 +18,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import static org.activiti.spring.conformance.set1.RuntimeTestConfiguration.*;
+import static org.activiti.spring.conformance.set1.Set1RuntimeTestConfiguration.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
