+++ activiti-spring-conformance/activiti-spring-conformance-util/pom.xml
@@ -9,16 +9,12 @@
     <version>7.0.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>activiti-spring-conformance-set2</artifactId>
+  <artifactId>activiti-spring-conformance-util</artifactId>
   <packaging>jar</packaging>
-  <name>Activiti Spring :: Conformance Set 2</name>
-  <description>Activiti Spring :: Conformance Set 2</description>
+  <name>Activiti Spring :: Conformance Util </name>
+  <description>Activiti Spring :: Conformance Util</description>
 
   <dependencies>
-    <dependency>
-      <groupId>org.activiti</groupId>
-      <artifactId>activiti-spring-conformance-util</artifactId>
-    </dependency>
     <dependency>
       <groupId>org.activiti.api</groupId>
       <artifactId>activiti-api-process-runtime</artifactId>
@@ -44,6 +40,10 @@
       <artifactId>activiti-api-model-shared</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.activiti.core.common</groupId>
+      <artifactId>activiti-spring-identity</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.activiti</groupId>
       <artifactId>activiti-engine</artifactId>
     </dependency>
@@ -60,6 +60,10 @@
       <artifactId>slf4j-api</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.assertj</groupId>
+      <artifactId>assertj-core</artifactId>
+    </dependency>
+    <dependency>
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
       <scope>test</scope>
@@ -67,7 +71,6 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
-      <scope>test</scope>
     </dependency>
   </dependencies>
