+++ activiti-spring-conformance/activiti-spring-conformance-util/src/main/java/org/activiti/spring/conformance/util/security/SecurityUtil.java
@@ -1,4 +1,4 @@
-package org.activiti.spring.conformance.set2.security.util;
+package org.activiti.spring.conformance.util.security;
 
 import org.activiti.api.runtime.shared.security.SecurityManager;
 import org.slf4j.Logger;
